+++ packages/interop/src/websites.spec.ts
@@ -1,33 +1,30 @@
 /* eslint-env mocha */
 import { createVerifiedFetch } from '@helia/verified-fetch'
 import { expect } from 'aegir/chai'
-import { createKuboNode } from './fixtures/create-kubo.js'
-import { loadFixtureDataCar } from './fixtures/load-fixture-data.js'
-import type { Controller } from 'ipfsd-ctl'
 
 describe('@helia/verified-fetch - websites', () => {
   describe('helia-identify.on.fleek.co', () => {
-    let controller: Controller<'go'>
     let verifiedFetch: Awaited<ReturnType<typeof createVerifiedFetch>>
 
     before(async () => {
-      controller = await createKuboNode()
-      await controller.start()
       // 2024-01-22 CID for _dnslink.helia-identify.on.fleek.co
-      await loadFixtureDataCar(controller, 'QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv-helia-identify-website.car')
       verifiedFetch = await createVerifiedFetch({
-        gateways: [`http://${controller.api.gatewayHost}:${controller.api.gatewayPort}`],
-        routers: [`http://${controller.api.gatewayHost}:${controller.api.gatewayPort}`]
+        gateways: ['http://127.0.0.1:8180'],
+        routers: ['http://127.0.0.1:8180'],
+        allowInsecure: true,
+        allowLocal: true
       })
     })
 
     after(async () => {
-      await controller.stop()
       await verifiedFetch.stop()
     })
 
     it('loads index.html when passed helia-identify.on.fleek.co root CID', async () => {
-      const resp = await verifiedFetch('ipfs://QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv')
+      const resp = await verifiedFetch('ipfs://QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv', {
+        allowLocal: true,
+        allowInsecure: true
+      })
       expect(resp).to.be.ok()
       const html = await resp.text()
       expect(html).to.be.ok()
@@ -35,7 +32,10 @@ describe('@helia/verified-fetch - websites', () => {
     })
 
     it('loads helia-identify.on.fleek.co index.html directly ', async () => {
-      const resp = await verifiedFetch('ipfs://QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv/index.html')
+      const resp = await verifiedFetch('ipfs://QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv/index.html', {
+        allowLocal: true,
+        allowInsecure: true
+      })
       expect(resp).to.be.ok()
       const html = await resp.text()
       expect(html).to.be.ok()
@@ -55,26 +55,26 @@ describe('@helia/verified-fetch - websites', () => {
    * ```
    */
   describe('fake blog.libp2p.io', () => {
-    let controller: Controller<'go'>
     let verifiedFetch: Awaited<ReturnType<typeof createVerifiedFetch>>
 
     before(async () => {
-      controller = await createKuboNode()
-      await controller.start()
-      await loadFixtureDataCar(controller, 'QmeiDMLtPUS3RT2xAcUwsNyZz169wPke2q7im9vZpVLSYw-fake-blog.libp2p.io.car')
       verifiedFetch = await createVerifiedFetch({
-        gateways: [`http://${controller.api.gatewayHost}:${controller.api.gatewayPort}`],
-        routers: [`http://${controller.api.gatewayHost}:${controller.api.gatewayPort}`]
+        gateways: ['http://127.0.0.1:8180'],
+        routers: ['http://127.0.0.1:8180'],
+        allowInsecure: true,
+        allowLocal: true
       })
     })
 
     after(async () => {
-      await controller.stop()
       await verifiedFetch.stop()
     })
 
     it('loads index.html when passed fake-blog.libp2p.io root CID', async () => {
-      const resp = await verifiedFetch('ipfs://QmeiDMLtPUS3RT2xAcUwsNyZz169wPke2q7im9vZpVLSYw')
+      const resp = await verifiedFetch('ipfs://QmeiDMLtPUS3RT2xAcUwsNyZz169wPke2q7im9vZpVLSYw', {
+        allowLocal: true,
+        allowInsecure: true
+      })
       expect(resp).to.be.ok()
       const html = await resp.text()
       expect(html).to.be.ok()